BUG: avoid overwriting output .qza files in table filtering tutorial
`make html` fails when running the commands in the table filtering tutorial because each command overwrites the output .qza from the previous command. The `commandblock` directive disallows output overwriting because only a single .qza link can be provided for each output .qza, making it impossible to link to each "version" of the .qza.