ENH: "Choose your own adventure"-style downloads
This adds support for download-method variations --- manual (browser),
`wget`, and `curl`. The aim here is for us to stop prescribing download
methods (this has lead to a lot of confusion in the past with different
versions of `curl`, as well as different `curl` configurations), and to
allow users to choose the best fit from a variety of options.

Secondarily, this circumvents some of those same `curl` problems, by
using `wget` when building the docs.