ENH: plugin directory supports multiple interfaces
The plugin directory now displays CLI and Artifact API help text for actions instead of a custom table of inputs, parameters, and outputs (this table was generally confusing to users).

Thanks @thermokarst for having the idea of using the help text that's already being generated for the CLI and Artifact API! The code is much simpler now too :)

Also, instead of using an action's underscored name (`Action.id`), the action's CLI name (dashed) is used in URLs and content.